requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.20.0
colorama==0.4.6
colorlog==6.8.0
contourpy==1.2.0
csvw==3.2.1
cycler==0.12.1
decorator==5.1.1
dlinfo==1.2.1
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
gradio==3.48.0
gradio_client==0.6.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
HyperPyYAML==1.2.2
idna==3.6
importlib-resources==6.1.1
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
julius==0.2.7
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.3
librosa==0.10.1
lightning==2.1.2
lightning-utilities==0.10.0
llvmlite==0.41.1
lxml==4.9.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
optuna==3.4.0
orjson==3.9.10
packaging==23.2
pandas==2.1.3
phonemizer==3.2.1
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
primePy==1.3
protobuf==4.25.1
pyannote.audio @ https://github.com/pyannote/pyannote-audio/archive/develop.zip
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pylatexenc==2.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.2
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.13.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
segments==2.2.1
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
speechtokenizer==0.1.2
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.1
torchmetrics==1.2.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
yarl==1.9.3